Logs plus clairs + corrections diverses
Le gros des modifications est dans plelib.sh.

Testé création :
    Standalone
    Dataguard
    RAC

oracle_preinstall/run_all.sh
    Le nom du paramètre est maintenant spécifié !
    Scripts impactés :
	database_servers/clone_master.sh

database_servers/install_grid.sh
	Correction code.

database_servers/install_oracle.sh
	Correction code.

db/create_db.sh
	Correction code.

db/create_pdb.sh
	Ajout du paramètre -from_pdb pour cloner à partir d'un pdb existant.

db/create_srv_for_rac_db.sh
	Correction bug administrator managed.

db/dbfs/create_crs_resource_for_dbfs.sh
	Correction bug, suite modification plelib.sh

db/drop_db.sh
	Amélioration code.

db/sql/adjust_recovery_size.sql
	La limite de la FRA passe de 80% à 90%

db/sql/lsusers.sql
	Évolutions.

db/wallet/create_wallet.sh
	Typo

db/wallet/drop_wallet.sh
	BUG si sqlnet.ora n'existe pas.

oracle_preinstall/01_create_oracle_users.sh
	BUG sur l'usage

oracle_preinstall/03_install_oracleasm.sh
	Typo

oracle_preinstall/04_apply_os_prerequis.sh
	BUG sur l'usage

plelib.sh
	Tente de rendre les logs plus clairs en remplaçants les chemins
	pas leur nom de variables.

san/check_target.sh
	Tente de rendre la sortie plus claire.

san/create_lun_for_db.sh
	Modification de l'usage

yum/backup_infra_repository.sh
	Affiche les informations sur le dernier backup.

db/sql/lsdiag.sql

+++ db/wallet/drop_wallet.sh
@@ -46,11 +46,21 @@ done
 
 must_be_user oracle
 
-info "Remove wallet : $wallet_path"
-LN
-execute_on_all_nodes_v2 'sed -i "/WALLET_LOCATION/d" $TNS_ADMIN/sqlnet.ora'
-execute_on_all_nodes_v2 'sed -i "/SQLNET.WALLET_OVERRIDE/d" $TNS_ADMIN/sqlnet.ora'
-LN
+if [ -f $TNS_ADMIN/sqlnet.ora ]
+then # sed plante si le fichier n'existe pas.
+	info "Remove wallet configuration."
+	execute_on_all_nodes_v2 'sed -i "/WALLET_LOCATION/d" $TNS_ADMIN/sqlnet.ora'
+	execute_on_all_nodes_v2 'sed -i "/SQLNET.WALLET_OVERRIDE/d" $TNS_ADMIN/sqlnet.ora'
+	LN
+else
+	info "\$TNS_ADMIN/sqlnet.ora not exists, nothing to do."
+fi
 
-execute_on_all_nodes "rm -rf $wallet_path"
-LN
+if [ -d "$wallet_path" ]
+then
+	info "$(replace_paths_by_shell_vars $wallet_path) exists : "
+	execute_on_all_nodes "rm -rf $wallet_path"
+	LN
+else
+	info "no wallet store exists."
+fi
